refactor: decouple Player specific systems from `systems.h`
