+++ src/ecs/entities.h
@@ -2,7 +2,7 @@
 
 #include "entity_builder.h"
 
-EntityBuilder PlayerCreate(f32 x, f32 y);
+#include "entities/player.h"
 EntityBuilder BlockCreate(f32 x, f32 y, f32 width, f32 height);
 EntityBuilder WalkerCreate(f32 x, f32 y);
 EntityBuilder CloudParticleCreate(f32 centerX, f32 centerY, Vector2 direction);
